汽车服务订单-额，懵逼了

+++ src/main/java/com/winstar/couponActivity/entity/JoinList.java
@@ -41,5 +41,5 @@ public class JoinList {
     private Date createTime;
 
     @Transient
-    private Integer number;
+    private long number;
 }
